requirements.txt
fastapi
httpx
langchain
pydantic
python-dotenv
sse-starlette
python-dotenv
uvicorn
tiktoken
openai
lark

pytz
faiss-cpu

pyppeteer
extractcontent3
beautifulsoup4
readability
google-api-python-client